sqlllm/requirements.txt
streamlit>=1.32
google-generativeai>=0.3.0
pandas>=2.0.0
python-dotenv>=0.19.0